+++ pushswap.h
@@ -6,7 +6,7 @@
 /*   By: thrio <thrio@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/14 09:30:59 by thrio             #+#    #+#             */
-/*   Updated: 2022/12/14 18:19:22 by thrio            ###   ########.fr       */
+/*   Updated: 2022/12/15 14:53:28 by thrio            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,11 +34,11 @@ int		ft_is_sorted(long *taba, int *size);
 int		ft_stack_len(int ac, char **av);
 char	**ft_split(char const *s, char c);
 int		ft_case_string(long *taba, int *size, char **av, char **buffer);
-int		ft_append_numbers(long *taba,char **av, int *size);
+int		ft_append_numbers(long *taba, char **av, int *size);
 int		ft_case_normal(long *taba, int *size, char **av);
 int		ft_is_numbers(char *numbers);
 int		ft_isdigit(int c);
-int		ft_atoi(const char *nptr);
+long	ft_atoi(const char *nptr);
 int		ft_is_repeated(long *taba, int size);
 int		ft_number_limits(long *taba, int size);
 void	rotate_a(long *taba, int size);
@@ -52,7 +52,7 @@ void	push_b(long *taba, long *tabb, int *size_a, int *size_b);
 void	reverse_b(long *tabb, int size);
 void	reverse_a(long *taba, int size);
 void	reverse_ab(long *taba, long *tabb, int size_a, int size_b);
-void    sort_three_digit(long *taba, int size);
-void    sort_five_digit(long *taba, long *tabb, int *size_a, int *size_b);
+void	sort_three_digit(long *taba, int size);
+void	sort_five_digit(long *taba, long *tabb, int *size_a, int *size_b);
 
 #endif
